Created enquire.py and moved the tests to enquire.py
