+++ river_core/main.py
@@ -1,14 +1,13 @@
 # See LICENSE for details
 """Console script for river_core."""
-
 import click
 import os
-
 from river_core.log import *
-from river_core.rivercore import rivercore_clean, rivercore_compile, rivercore_generate, rivercore_merge, rivercore_setup, rivercore_enquire
+from river_core.rivercore import rivercore_clean, rivercore_compile, rivercore_generate, rivercore_merge, rivercore_setup
 from river_core.__init__ import __version__
 import river_core.constants as constants
-
+import river_core.utils as utils
+import pytest
 
 def check_config():
     """ Checks if 
@@ -188,7 +187,11 @@ def compile(config, test_list, coverage, verbosity, dut_stage, ref_stage,
               required=True)
 @cli.command()
 def enquire(test_list):
-    rivercore_enquire(test_list)
+    enquire.test_list = test_list
+    pytest.main(['--log-cli-level=0', \
+                 '--html=test_enquire-report.html', \
+                '--self-contained-html', \
+                __file__.rstrip('main.py')+'enquire.py'])
 
 
 @click.version_option(version=__version__)
