+++ river_core/rivercore.py
@@ -1125,75 +1125,4 @@ def rivercore_setup(config, dut, gen, ref, verbosity):
         logger.info(
             'Created {0} Plugin in the current working directory'.format(ref))
 
-
-def rivercore_enquire(testyaml):
-    testyaml_dict = utils.load_yaml(testyaml)
-    #@pytest.mark.parametrize('testname', testyaml_dict)
-    def test_enquire(testname):
-        '''
-        Utility function that gives the status of each test 
-        '''
-        node = testyaml_dict[testname]
-        # check if compilation of the test is over
-        elffile = node['work_dir'] + '/'+testname + '.elf'
-        if not os.path.exists(elffile):
-            assert False, testname + ' has not compiled.\n'
-        else:
-            dumpfile = node['work_dir'] + '/dut.dump'
-            if os.path.exists(dumpfile):
-                # check if dtl.dump is created
-                # check if dut.dump is complete
-                binary = lief.parse(elffile)
-                tohost_addr = str.format('0x{:08X}', int(hex(binary.get_symbol('tohost').value), 16))
-                with open(dumpfile) as rtlfptr:
-                    dump_lines = rtlfptr.readlines()
-                    last_line = dump_lines[-1]
-                    if tohost_addr not in last_line:
-                        assert False, testname + ' tohost is not written to yet'
-                    else:
-                        # check if spike simulation is over
-                        spikedumpfile = node['work_dir'] + '/temp'
-                        if not os.path.exists(spikedumpfile):
-                            assert False, testname + ' spike has not finished execution'
-                        else:
-                            # check if tohost is written to in spike.dump
-                            with open(spikedumpfile) as spikefptr:
-                                spike_dump_lines = spikefptr.readlines()
-                                last_line = spike_dump_lines[-1]
-                                if tohost_addr not in last_line:
-                                    assert False, testname + ' spike simulation has some errors'
-            else:
-                dumpfile = node['work_dir'] + '/rtl'+'0'+'.dump'
-                if not os.path.exists(dumpfile):
-                    assert False, testname + ' rtl dump not created'
-                # check if rtl<coreid>.dump is complete
-                else:
-                    binary = lief.parse(elffile)
-                    tohost_addr = str.format('0x{:08X}', int(hex(binary.get_symbol('tohost').value), 16))
-                    with open(dumpfile) as rtlfptr:
-                        dump_lines = rtlfptr.readlines()
-                        last_line = dump_lines[-1]
-                        if tohost_addr not in last_line:
-                            assert False, testname + ' tohost is not written to yet'
-                        else:
-                            # check if spike simulation is over
-                            spikedumpfile = node['work_dir'] + '/temp'
-                            if not os.path.exists(spikedumpfile):
-                                assert False, testname + ' spike has not finished execution'
-                            else:
-                                # check if tohost is written to in spike.dump
-                                with open(spikedumpfile) as spikefptr:
-                                    spike_dump_lines = spikefptr.readlines()
-                                    last_line = spike_dump_lines[-1]
-                                    if tohost_addr not in last_line:
-                                        assert False, testname + ' spike simulation has some errors'
-    for testname in testyaml_dict.keys():
-        test_enquire(testname)
-    # cmd = 'pytest -k enquire \
-    #     -o log_cli=True --capture=tee-sys -v --log-cli-level=0 \
-    #     --html=test_enquire-report.html --self-contained-html\
-    #      --noconftest'
-    # errcode, rout, rerr = utils.sys_command(cmd, logging =False)
-    # print(errcode)
-    # print(rout)
-    # print(rerr)
+
